pipeline/pyproject.toml
[project]
name = "trump-prediction-pipeline"
version = "0.1.0"
description = "Automated evaluation pipeline for LLM leader prediction — Bosse et al. (2026) architecture adapted for Trump decisions"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.4",
    "langchain-core>=0.3",
    "langchain-google-genai>=4.0",
    "langchain-anthropic>=0.3",
    "langchain-tavily>=0.2",
    "pydantic>=2.0",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "python-dotenv>=1.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
